Don't describe constants
This means they don't have to exist when running the entire test suite,
which is beneficial because it has students writing the classes
themself, which is a valuable habit to get into.

The file existing needs to happen for `require_relative` to work, which
seems like a fine concession for now.